refactor: move channels_v2/facebook_channel.py to channels/facebook_channel.py
