+++ apps/channels/tasks.py
@@ -9,7 +9,6 @@ from twilio.request_validator import RequestValidator
 
 from apps.channels.api_channel import ApiChannel
 from apps.channels.channels_v2.evaluation_channel import EvaluationChannel
-from apps.channels.channels_v2.facebook_channel import FacebookMessengerChannel
 from apps.channels.channels_v2.sureadhere_channel import SureAdhereChannel
 from apps.channels.channels_v2.telegram_channel import TelegramChannel
 from apps.channels.channels_v2.whatsapp_channel import WhatsappChannel
@@ -23,6 +22,7 @@ from apps.channels.datamodels import (
     WhatsAppMessage,
 )
 from apps.channels.datamodels import EmailMessage as EmailMessageDatamodel
+from apps.channels.facebook_channel import FacebookMessengerChannel
 from apps.channels.models import ChannelPlatform, ExperimentChannel
 from apps.chat.models import ChatMessage
 from apps.chatbots.version_resolver import resolve_published_or_working
